# blinddatingapp-frontend/src/generated-sources/models/Gender.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Profile Service API
 * Profile Service api definitie
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Gender = {
    Man: 'MAN',
    Woman: 'WOMAN',
    X: 'X'
} as const;
export type Gender = typeof Gender[keyof typeof Gender];


export function instanceOfGender(value: any): boolean {
    for (const key in Gender) {
        if (Object.prototype.hasOwnProperty.call(Gender, key)) {
            if (Gender[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function GenderFromJSON(json: any): Gender {
    return GenderFromJSONTyped(json, false);
}

export function GenderFromJSONTyped(json: any, ignoreDiscriminator: boolean): Gender {
    return json as Gender;
}

export function GenderToJSON(value?: Gender | null): any {
    return value as any;
}