# blinddatingapp-frontend/src/generated-sources/models/HairColor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Profile Service API
 * Profile Service api definitie
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const HairColor = {
    Blond: 'BLOND',
    Brown: 'BROWN',
    Black: 'BLACK',
    Red: 'RED',
    False: 'false',
    Gray: 'GRAY',
    Pink: 'PINK'
} as const;
export type HairColor = typeof HairColor[keyof typeof HairColor];


export function instanceOfHairColor(value: any): boolean {
    for (const key in HairColor) {
        if (Object.prototype.hasOwnProperty.call(HairColor, key)) {
            if (HairColor[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function HairColorFromJSON(json: any): HairColor {
    return HairColorFromJSONTyped(json, false);
}

export function HairColorFromJSONTyped(json: any, ignoreDiscriminator: boolean): HairColor {
    return json as HairColor;
}

export function HairColorToJSON(value?: HairColor | null): any {
    return value as any;
}