# blinddatingapp-frontend/src/generated-sources/models/MusicGenre.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Profile Service API
 * Profile Service api definitie
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const MusicGenre = {
    Rock: 'ROCK',
    Metal: 'METAL',
    Deathmetal: 'DEATHMETAL',
    Trashmetal: 'TRASHMETAL',
    Ballroom: 'BALLROOM',
    Classical: 'CLASSICAL',
    Country: 'COUNTRY',
    Dance: 'DANCE',
    Electronic: 'ELECTRONIC',
    Funk: 'FUNK',
    HipHop: 'HIP_HOP',
    Jazz: 'JAZZ',
    Latin: 'LATIN',
    Pop: 'POP',
    Rb: 'RB',
    Reggae: 'REGGAE',
    Unknown: 'UNKNOWN',
    World: 'WORLD'
} as const;
export type MusicGenre = typeof MusicGenre[keyof typeof MusicGenre];


export function instanceOfMusicGenre(value: any): boolean {
    for (const key in MusicGenre) {
        if (Object.prototype.hasOwnProperty.call(MusicGenre, key)) {
            if (MusicGenre[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function MusicGenreFromJSON(json: any): MusicGenre {
    return MusicGenreFromJSONTyped(json, false);
}

export function MusicGenreFromJSONTyped(json: any, ignoreDiscriminator: boolean): MusicGenre {
    return json as MusicGenre;
}

export function MusicGenreToJSON(value?: MusicGenre | null): any {
    return value as any;
}